R Interface for

Pulls posts

posts_search(
  token,
  accountTypes = NULL,
  and = NULL,
  brandedContent = "no_filter",
  count = 10,
  endDate = ct_now(),
  includeHistory = NULL,
  inAccountIds = NULL,
  inListIds = NULL,
  language = NULL,
  minInteractions = 0,
  minSubscriberCount = 0,
  not = NULL,
  notInAccountIds = NULL,
  notInLstIds = NULL,
  notInTitle = NULL,
  offset = 0,
  pageAdminTopCountry = NULL,
  platforms = NULL,
  searchField = "text_fields_and_image_text",
  searchTerm = "",
  startDate = NULL,
  sortBy = "date",
  timeframe = NULL,
  types = NULL,
  verified = "no_filter",
  verifiedOnly = "false",
  search10k = FALSE,
  boolean_allowed = FALSE,
  output_raw = TRUE
)

Arguments

token
A valid CT API Token

accountTypes
...

and
...

brandedContent
...

count
...

endDate
...

includeHistory
...

inAccountIds
...

inListIds
...

language
...

minInteractions
...

minSubscriberCount
...

not
...

notInAccountIds
...

notInLstIds
...

notInTitle
...

offset
...

pageAdminTopCountry
...

platforms
...

searchField
...

searchTerm
...

startDate
...

sortBy
...

timeframe
...

types
...

verified
...

verifiedOnly
...

search10k
A boolean signifying whether you have elavated access for higher count numbers.

boolean_allowed
A boolean signifying whether you have boolean term search enabled.

output_raw
A boolean signifying whether to output raw CT output or parse into df. Defaulting to TRUE.

Examples

# Pull all posts from November 2020 by CNN containing the term "Trump"
out <- posts_search(token = crowdtangle_token(),
                    inAccountIds = "8323", #CNN example
                    searchTerm = "Trump",
                    startDate = "2020-11-01T00:00:00",
                    endDate = "2020-11-30T23:59:59",
                    count = 10000,
                    sortBy = 'date',
                    platforms = 'facebook',
                    search10k = T,
                    boolean_allowed = T,
                    output_raw = TRUE)

#> [1] "collecting posts"
#> [1] "try for the 1th time."
#> [1] "now collected 256 entries"